Reorganize the python package
