+++ clients/client-python/gravitino/client/generic_model_version.py
@@ -16,7 +16,7 @@
 # under the License.
 from typing import Optional, Dict, List
 
-from gravitino.api.model_version import ModelVersion
+from gravitino.api.model.model_version import ModelVersion
 from gravitino.dto.audit_dto import AuditDTO
 from gravitino.dto.model_version_dto import ModelVersionDTO
