+++ clients/client-python/tests/unittests/test_model_catalog_api.py
@@ -20,8 +20,8 @@ from http.client import HTTPResponse
 from unittest.mock import Mock, patch
 
 from gravitino import NameIdentifier, GravitinoClient
-from gravitino.api.model import Model
-from gravitino.api.model_version import ModelVersion
+from gravitino.api.model.model import Model
+from gravitino.api.model.model_version import ModelVersion
 from gravitino.dto.audit_dto import AuditDTO
 from gravitino.dto.model_dto import ModelDTO
 from gravitino.dto.model_version_dto import ModelVersionDTO
